+++ jib-core/src/main/java/com/google/cloud/tools/jib/filesystem/DirectoryWalker.java
@@ -90,11 +90,7 @@ public class DirectoryWalker {
    */
   public ImmutableList<Path> walk() throws IOException {
     try (Stream<Path> fileStream = Files.walk(rootDir)) {
-      Stream<Path> filteredFileStream = fileStream;
-      if (pathFilter != null) {
-        filteredFileStream = fileStream.filter(pathFilter);
-      }
-      return filteredFileStream.sorted().collect(ImmutableList.toImmutableList());
+      return fileStream.filter(pathFilter).sorted().collect(ImmutableList.toImmutableList());
     }
   }
 }