+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildImageTask.java
@@ -79,8 +79,8 @@ public class BuildImageTask extends DefaultTask {
   public void buildImage() throws InvalidImageReferenceException {
     // Asserts required @Input parameters are not null.
     Preconditions.checkNotNull(jibExtension);
-    GradleBuildLogger gradleBuildLogger = new GradleBuildLogger(getLogger());
-    jibExtension.handleDeprecatedParameters(gradleBuildLogger);
+    GradleJibLogger gradleJibLogger = new GradleJibLogger(getLogger());
+    jibExtension.handleDeprecatedParameters(gradleJibLogger);
     SystemPropertyValidator.checkHttpTimeoutProperty(GradleException::new);
 
     if (Strings.isNullOrEmpty(jibExtension.getTargetImage())) {
@@ -91,7 +91,7 @@ public class BuildImageTask extends DefaultTask {
     }
 
     if (Boolean.getBoolean("sendCredentialsOverHttp")) {
-      gradleBuildLogger.warn(
+      gradleJibLogger.warn(
           "Authentication over HTTP is enabled. It is strongly recommended that you do not enable "
               + "this on a public network!");
     }
@@ -108,7 +108,7 @@ public class BuildImageTask extends DefaultTask {
 
     GradleProjectProperties gradleProjectProperties =
         GradleProjectProperties.getForProject(
-            getProject(), gradleBuildLogger, jibExtension.getExtraDirectoryPath());
+            getProject(), gradleJibLogger, jibExtension.getExtraDirectoryPath());
     String mainClass = gradleProjectProperties.getMainClass(jibExtension);
 
     // Builds the BuildConfiguration.
@@ -132,13 +132,13 @@ public class BuildImageTask extends DefaultTask {
             .setProgramArguments(jibExtension.getArgs())
             .setExposedPorts(ExposedPortsParser.parse(jibExtension.getExposedPorts()));
     if (jibExtension.getUseCurrentTimestamp()) {
-      gradleBuildLogger.warn(
+      gradleJibLogger.warn(
           "Setting image creation time to current time; your image may not be reproducible.");
       containerConfigurationBuilder.setCreationTime(Instant.now());
     }
 
     BuildConfiguration.Builder buildConfigurationBuilder =
-        BuildConfiguration.builder(gradleBuildLogger)
+        BuildConfiguration.builder(gradleJibLogger)
             .setBaseImageConfiguration(baseImageConfiguration)
             .setTargetImageConfiguration(targetImageConfiguration)
             .setContainerConfiguration(containerConfigurationBuilder.build())
@@ -158,7 +158,7 @@ public class BuildImageTask extends DefaultTask {
     BuildConfiguration buildConfiguration = buildConfigurationBuilder.build();
 
     // TODO: Instead of disabling logging, have authentication credentials be provided
-    GradleBuildLogger.disableHttpLogging();
+    GradleJibLogger.disableHttpLogging();
 
     RegistryClient.setUserAgentSuffix(USER_AGENT_SUFFIX);